Description simplified!!! all the data will be in the log!!!!

+++ src/core/res_repo_dir_structure/res_ider/builder.jl
@@ -1,4 +1,4 @@
 # A res identifier (res_ider) is composed by a custom name (res_name)
 # and a time generated tag (time_tag)
-build_res_ider(res_name::String, time_tag::Int128) = string(validate_res_name(res_name),
+build_res_ider(res_name::String, time_tag::Int) = string(validate_res_name(res_name),
     res_name_time_tag_separator, time_tag);
